fix: fikset problem med at samme pris fra samme butikk ble lagt til flere ganger
